requirements.txt
colorama~=0.4.6
PyYAML~=6.0.2
jsonpath-ng~=1.7.0
playwright~=1.53.0
pytest-playwright~=0.7.0
pandas~=2.3.1
responses~=0.25.7
requests~=2.32.4
pytest~=8.4.1
Faker~=37.5.3
jsonschema~=4.25.1